Extract format-agnostic monthly aggregation from the netCDF writer
monthly.py mixed two responsibilities: aggregating batches into
per-calendar-month running means, and representing the result as netCDF.
Move the first into monthly_aggregation.py — the month indexing (now a
MonthIndexer object rather than two writer attributes and two methods),
the running-mean update, the valid-time calendar arithmetic, and the
dimension and variable names those are stored under. MonthlyDataWriter
keeps only the netCDF representation.

No behavior change.
